+++ RELEASE_NOTES.md
@@ -9,7 +9,7 @@ permalink: /docs/release-notes
 
 # Release Notes
 
-## 2.3.0-beta1
+## 2.3.0
 
 * Synergize no longer uses Bonjour/Zeroconf to discover the Synergia
   virtual instrument. A much more reliable mechanism has been
@@ -51,6 +51,10 @@ permalink: /docs/release-notes
       error message.
     * Fixes [issue #20](https://github.com/chinenual/synergize/issues/20):
       Mute and solo buttons did not transmit to control surface.
+* Fixes [issue #27](https://github.com/chinenual/synergize/issues/27):
+  When in Voicing Mode, disconnecting the Synergy left the Synergize
+  side editor in voicing mode.  Now disconnecting the Synergy
+  implicitly turns off voicing mode.
   
 ## 2.2.0